sparse matrix added
patterns hold an array of 0s and 1s that corresponds to the entity map.
0 says that the corresponding entity is not present while a 1 says that
it is. When building the simTable this saves performing as many loops
which speeds up the process immensely (4000 patterns in 26 seconds).
Also added the date variable to time how long the process takes.
Finally when building the simTable the j iterator is set to the value
of the i iterator. This saves comparing the same two patterns twice
(effectively halting the time it takes to perform the function).